Dirty hack to get instruction execution profiles
The NVidia Profiler will only show instruction execution profiles
for the PTX if the source PTX included some line number
information. Here we messily insert into the generated PTX the
source file and starting line of the function that was jitted, in
order to get the instruction profile to show up, and make it
display the source of the jitted function.

Note that this is the "wrong" way to go about this in general - a
nice approach would be to insert debug info into the LLVM IR as
specified in the NVVM IR specification, which would be much
cleaner and also allow tying back PTX instructions to specific
Python source lines.