WIP on thread rank
Was starting to add implementation of thread rank. Noticed that there
seems to be some extraneous stuff in cuda target - committing here to
save progess whilst switching to another branch to tidy up first. Some
additions so far on this branch are also not needed and could do with
cutting down.