CUDA: Require CUDA 9.0
The `syncwarp()` function used in the reduction kernels was introduced
with CUDA 9.0. Rather than trying to make the availability of reduction
kernels conditional on the CUDA Toolkit version, it seems more sensible
to just drop CUDA 8.0 now that it is quite old.

This commit also includes fixes for the logic setting `SUPPORTED_CC` in
`nvvm.py` (when there was no suitable toolkit, it was creating a tuple
containing the empty tuple) and added an error when an attempt is made
to use NVVM when no supported toolkit is found.