Refactor so that definitions are no longer keyed by CC
This means that it should now be the same as self.overloads - the aim is
to remove it in favour of self.overloads, used by the dispatcher.

Some modifications of inspect_asm, inspect_llvm, inspect_sass were
necessary - there were some oddities about these methods in that
specifying the compute capability actually had no effect on some of
them - LLVM is device independent, and SASS was produced for the current
device regardless of the CC requested.