Make Dispatcher.compile() work for external calls
This doesn't result in any new passes but does help some tests get
further. THe next problem to solve is the discrepancy in the types in
overloads, which have the launch args appended, and what functions
expect to pass to inspect_llvm() etc... probably they should normalize
the args too.