Add more stuff for debugging the issue
Seems like the version with a type change does not automatically get
linked in. possibly this works on the CPU target because the
non-recursively called function was jitted and is available in the
symbol table, so is not an unresolved call. Maybe we need to scan for
unresolved calls in another way. Check:

- Does the version with uid 7 still appear in the list of known
  references?
- Can we scan the IR for unresolved references, find the call to the
  version with uid 7, and link it?
- Does scanning accidentally find / attempt to link other things that we
  should not attempt to resolve, e.g. libdevice functions? If so, can we
  possibly avoid attempting to link them because they're not things we
  know a definition of?