A couple of cheeky hacks to get successful jit on riscv
We disable optimisation (because some segfault occurs during opt passes, maybe related
to the fact I haven't properly updated llvmlite for llvm 15). This causes some issue for NRT
so I've disabled that as well. That seems to be enough to jit a function and call it successfully
on riscv, so that the function

```
@njit
def f(a, b):
    return a + b
```

can be successfully compiled and called so that `f(1, 2)` returns 3.weq